Exclude signature directory from compressible objects
It seems better not to compress signature directories.
Various PDF digitally signing tools, including Adobe Acrobat Reader DC,
do not compress signature directories.

Table 8.93 "Entries in a signature dictionary" in PDF 1.5 reference
describes that /ByteRange in the signature directory
shall be used to describe a digest
that does not include the signature value (/Contents) itself.

The byte ranges cannot be determined if the directory is compressed.